requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.3
langchain-text-splitters==0.3.2
langchain-groq==0.2.1
langchain-huggingface==0.1.2
chromadb==0.5.5
sentence-transformers==3.1.1
python-dotenv==1.0.1
huggingface-hub>=0.24.5
torch>=2.1.0
transformers>=4.40.0
python-docx==1.1.2